services/gateway/api-gateway/pyproject.toml
[project]
name = "temuhash-api-gateway"
version = "0.1.0"
description = "API gateway for TemuHash marketplace microservices"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.1",
  "httpx>=0.27.0",
  "python-jose[cryptography]>=3.3.0",
  "pydantic-settings>=2.3.0",
  "grpcio>=1.64.0",
  "grpcio-tools>=1.64.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-exporter-otlp>=1.25.0"
]

[tool.uv]
index-url = "https://pypi.org/simple"
